TAVERNA-963 package taverna.externaltool.*
