+++ taverna-external-tool-activity/src/main/java/org/apache/taverna/activities/externaltool/desc/ScriptInput.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package de.uni_luebeck.inb.knowarc.usecases;
+package org.apache.taverna.activities.externaltool.desc;
 
 import java.nio.charset.Charset;
