+++ taverna-external-tool-activity/src/main/java/org/apache/taverna/activities/externaltool/desc/ScriptOutput.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package de.uni_luebeck.inb.knowarc.usecases;
+package org.apache.taverna.activities.externaltool.desc;
 import java.util.ArrayList;
 
 import org.apache.taverna.activities.externaltool.ExternalToolActivity;
@@ -24,59 +24,48 @@ import org.apache.taverna.workflowmodel.processor.config.ConfigurationBean;
 import org.apache.taverna.workflowmodel.processor.config.ConfigurationProperty;
 
 /**
- * Internal description of output
+ * Internal description of input
  */
-@ConfigurationBean(uri = ExternalToolActivity.URI + "#ScriptInput")
-public class ScriptInputUser extends ScriptInput {
-
-	/**
-	 * This input may be fed from multiple ouputs.
-	 */
-	private boolean list = false;
-	/**
-	 * True if the data from a list input in taverna is concatenated into one single input file.
-	 */
-	private boolean concatenate = false;
-
+@ConfigurationBean(uri = ExternalToolActivity.URI + "#ScriptOutput")
+public class ScriptOutput {
+	private String path;
+	private boolean binary;
 	private ArrayList<String> mime = new ArrayList<String>();
 
 	@Override
 	public String toString() {
-		return "Input[tag: " + getTag() + (isFile() ? ", file" : "")
-				+ (isTempFile() ? ", tempfile" : "")
-				+ (isBinary() ? ", binary" : "") + (list ? ", list" : "")
-				+ (concatenate ? ", concatenate" : "")
+		return "Output[path: " + path + (binary ? ", binary" : "")
 				+ " mime: " + mime.toString() + "]";
 	}
 
 	/**
-	 * @return the list
+	 * @return the path
 	 */
-	public final boolean isList() {
-		return list;
+	public final String getPath() {
+		return path;
 	}
 
 	/**
-	 * @param list the list to set
+	 * @param path the path to set
 	 */
-	@ConfigurationProperty(name = "list", label = "List")
-	public final void setList(boolean list) {
-		this.list = list;
+	@ConfigurationProperty(name = "path", label = "Path")
+	public final void setPath(String path) {
+		this.path = path;
 	}
 
 	/**
-	 * @return the concatenate
+	 * @return the binary
 	 */
-	public final boolean isConcatenate() {
-		return concatenate;
+	public final boolean isBinary() {
+		return binary;
 	}
 
 	/**
-	 * @param concatenate the concatenate to set
+	 * @param binary the binary to set
 	 */
-	@ConfigurationProperty(name = "concatenate", label = "Concatenate")
-	public final void setConcatenate(boolean concatenate) {
-		this.concatenate = concatenate;
+	@ConfigurationProperty(name = "binary", label = "Binary")
+	public final void setBinary(boolean binary) {
+		this.binary = binary;
 	}
 
 	/**
