+++ taverna-external-tool-activity/src/main/java/org/apache/taverna/activities/externaltool/invocation/AskUserForPw.java
@@ -16,21 +16,17 @@
  * limitations under the License.
  */
 
-package de.uni_luebeck.inb.knowarc.usecases.invocation;
+package org.apache.taverna.activities.externaltool.invocation;
 
-public class InvocationException extends Exception {
+public interface AskUserForPw {
 
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 437316164959631591L;
+	// we will try keyfile first
+	public String getKeyfile();
+	public String getPassphrase();
 
-	public InvocationException(String string) {
-		super(string);
-	}
-	
-	public InvocationException(Exception e) {
-		super(e);
-	}
+	// then username+password
+	public String getUsername();
+	public String getPassword();
 
+	public void authenticationSucceeded();
 }
