+++ taverna-external-tool-activity/src/main/resources/META-INF/services/org.apache.taverna.reference.ExternalReferenceSPI
@@ -1,2 +1,2 @@
 # Implementation classes of ExternalReferenceSPI go here, one per line
-de.uni_luebeck.inb.knowarc.usecases.invocation.ssh.SshReference
+org.apache.taverna.activities.externaltool.invocation.ssh.SshReference
