+++ taverna-external-tool-activity/src/main/resources/META-INF/services/org.apache.taverna.reference.ValueToReferenceConverterSPI
@@ -1,3 +1,3 @@
 # Implementation classes of ValueToReferenceConverterSPI go here, one per line
-de.uni_luebeck.inb.knowarc.usecases.invocation.ssh.SshUrlToSshReference
+org.apache.taverna.activities.externaltool.invocation.ssh.SshUrlToSshReference
